+++ Blocks5/src/cf_camera.cpp
@@ -9,20 +9,20 @@ CF_Camera::~CF_Camera()
 {
 }
 
-void CF_Camera::render(double t,
+void CF_Camera::render(float t,
 					   uint oldImageID,
 					   uint newImageID)
 {
 	// The camera looks along the y axis, from the old image up to the new
 	// one four units above it.
-	const Mat4 projection = Mat4::perspective(90.0, 1.0, 0.1, 100.0);
+	const Mat4 projection = Mat4::perspective(90.0f, 1.0f, 0.1f, 100.0f);
 
-	double s1 = sin(t * 1.5707963267948966192313216916398);
-	double s2 = sin(t * 3.1415926535897932384626433832795);
-	double y = s1 * s1 * s1 * s1 * s1 * s1 * 4.0;
-	double l = y + s2 * s2 * s2 * s2 * 4.0;
+	float s1 = sinf(t * 1.5707963267948966192313216916398f);
+	float s2 = sinf(t * 3.1415926535897932384626433832795f);
+	float y = s1 * s1 * s1 * s1 * s1 * s1 * 4.0f;
+	float l = y + s2 * s2 * s2 * s2 * 4.0f;
 
-	const Mat4 transform = projection * Mat4::lookAt(0.0, y, -1.0, 0.0, l, 0.0, 0.0, 1.0, 0.0);
+	const Mat4 transform = projection * Mat4::lookAt(0.0f, y, -1.0f, 0.0f, l, 0.0f, 0.0f, 1.0f, 0.0f);
 
 	Renderer::inst().clear(Vec4f(0.0f, 0.0f, 0.0f, 1.0f));
