+++ Blocks5/src/cf_cube.h
@@ -11,7 +11,7 @@ public:
 	CF_Cube();
 	~CF_Cube();
 
-	void render(double t, uint oldImageID, uint newImageID);
+	void render(float t, uint oldImageID, uint newImageID);
 };
 
 #endif
