+++ Blocks5/src/exit.h
@@ -12,7 +12,7 @@ public:
 	~Exit();
 
 	void onRemove();
-	void onRender(RenderLayer layer, const Vec4d& color);
+	void onRender(RenderLayer layer, const Vec4f& color);
 	void updateSprites();
 	void onUpdate();
 };
