+++ Blocks5/src/gui_checkbox.cpp
@@ -28,7 +28,7 @@ void GUI_CheckBox::onRender()
 		{
 			// draw the checkmark
 			Vec2i offset = (size - Vec2i(16, 16)) / 2;
-			Engine::inst().renderSprite(gui.getSkin(), offset, Vec2i(32, 224), Vec2i(16, 16), Vec4d(1.0));
+			Engine::inst().renderSprite(gui.getSkin(), offset, Vec2i(32, 224), Vec2i(16, 16), Vec4f(1.0f));
 		}
 	}
 	else
@@ -60,7 +60,7 @@ void GUI_CheckBox::onRender()
 	Vec2i dim;
 	std::string title = localizeString(this->title);
 	p_font->measureText(title, &dim, 0);
-	p_font->renderText(title, Vec2i(size.x + 10, (size.y - dim.y) / 2), active ? Vec4d(1.0, 1.0, 1.0, 1.0) : Vec4d(0.5, 0.5, 0.5, 1.0));
+	p_font->renderText(title, Vec2i(size.x + 10, (size.y - dim.y) / 2), active ? Vec4f(1.0f, 1.0f, 1.0f, 1.0f) : Vec4f(0.5f, 0.5f, 0.5f, 1.0f));
 }
 
 void GUI_CheckBox::onMouseDown(const Vec2i& position,
